+++ cmake/NeuronFileLists.cmake
@@ -47,7 +47,6 @@ set(HEADER_FILES_TO_INSTALL
     ocmisc.h
     options.h
     parse_with_deps.hpp
-    random1_fwd.h
     scoplib.h
     section.h
     spconfig.h
