+++ src/oc/nrnrandom.h
@@ -1,8 +1,7 @@
 #pragma once
-#include "random1_fwd.h"
-
 #include <stdint.h>
 
+class Rand;
 long nrn_get_random_sequence(Rand* r);
 Rand* nrn_random_arg(int);
 int nrn_random_isran123(Rand* r, uint32_t* id1, uint32_t* id2, uint32_t* id3);
